pyproject.toml
[project]
name = "rlane-kroger"
version = "1.0.0"
description = "Tools for Kroger MyInfo and MyTime apps."
authors = [
    {name = "Russel Lane", email = "russel@rlane.com"},
]
license = {text = "MIT"}
readme = "README.md"
keywords = ["kroger", "python"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: General",
]
requires-python = ">=3.10"
dependencies = [
    "rlane-libcli>=1.0.3",
    "pdfminer-six>=20231228",
    "selenium>=4.21.0",
]

[project.urls]
Homepage = "https://github.com/russellane/kroger"

[project.scripts]
kroger = 'kroger.cli:main'

[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.4.2",
    "flake8-bugbear>=24.4.26",
    "flake8-pydocstyle>=0.2.4",
    "flake8-pylint>=0.2.1",
    "flake8-pytest-style>=2.0.0",
    "flake8-simplify>=0.21.0",
    "flake8>=7.0.0",
    "isort>=5.13.2",
    "pytest-cov>=5.0.0",
    "pytest>=8.2.1",
    "toml>=0.10.2",
]

[tool.black]
line-length = "97"

[tool.isort]
line_length = "97"
profile = "black"

[tool.pylint.format]
max-line-length = "97"

[tool.pylint.messages_control]
enable = [
    "useless-suppression",
]
fail-on = [
    # useless-suppression
    "I0021",
]

[tool.pydocstyle]
convention = "google"
add-ignore = [
    # "D105",     # pep257 numpy google   # Missing docstring in magic method
    "D105",
    # "D202",     # pep257 numpy google   # No blank lines allowed after function docstring
    "D202",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

# vim: set ts=4 sw=4 et:
